refactor l10n -> loca reenable "monkey" (requires `-gcflags=-l` on test run!)
