+++ .travis.yml
@@ -23,7 +23,7 @@ before_install:
 script:
   - golint ./...
   - go vet ./...
-  - go test -covermode=count -coverprofile=profile.cov ./...
+  - go test -gcflags=-l -covermode=count -coverprofile=profile.cov ./...
   - goveralls -coverprofile=profile.cov -service=travis-ci
 
 before_deploy:
