+++ loca/loca.go
@@ -1,0 +1,28 @@
+package loca
+
+import (
+	"github.com/drpsychick/alexa-go-cloudformation-demo/pkg/l10n"
+)
+
+const (
+	GreetingTitle     l10n.Key = "greeting_title"
+	Greeting          l10n.Key = "greeting"
+	GreetingSSML      l10n.Key = "greeting_ssml"
+	ByeBye            l10n.Key = "byebye"
+	StopTitle         l10n.Key = "stop_title"
+	Stop              l10n.Key = "stop"
+	SaySomethingTitle l10n.Key = "saysomething_title"
+	SaySomething      l10n.Key = "saysomething"
+	SaySomethingSSML  l10n.Key = "saysomething_ssml"
+)
+
+func init() {
+	var locales = []*l10n.Locale{
+		deDE, enUS, frFR,
+	}
+	for _, l := range locales {
+		if err := l10n.Register(l); err != nil {
+			panic("registration of locale failed")
+		}
+	}
+}
